feat: add part of main service

+++ main-service/src/main/java/ru/practicum/mainservice/categories/service/CategoriesServiceImpl.java
@@ -9,6 +9,10 @@ import ru.practicum.mainservice.categories.dto.NewCategoryDto;
 import ru.practicum.mainservice.categories.entity.CategoriesEntity;
 import ru.practicum.mainservice.categories.mapper.CategoriesMapper;
 import ru.practicum.mainservice.categories.repository.CategoriesRepository;
+import ru.practicum.mainservice.errors.exceptions.DataNotFoundException;
+import ru.practicum.mainservice.errors.exceptions.EventValidationException;
+import ru.practicum.mainservice.events.entity.EventsEntity;
+import ru.practicum.mainservice.events.repository.EventsRepository;
 import ru.practicum.mainservice.users.dto.NewUserRequest;
 import ru.practicum.mainservice.users.dto.UserDto;
 import ru.practicum.mainservice.users.entity.UserEntity;
@@ -20,6 +24,7 @@ import java.util.List;
 public class CategoriesServiceImpl implements CategoriesService {
     private final CategoriesRepository repository;
     private final CategoriesMapper mapper;
+    private final EventsRepository eventsRepository;
 
     @Override
     public CategoryDto addCategory(NewCategoryDto request) {
@@ -30,6 +35,9 @@ public class CategoriesServiceImpl implements CategoriesService {
 
     @Override
     public CategoryDto updateCategory(Integer catId, NewCategoryDto request) {
+        if (repository.existsById(catId)) {
+            throw new DataNotFoundException("Category with id=" + catId + " was not found");
+        }
         CategoriesEntity entity = findCategoryById(catId);
         entity.setName(request.getName());
         CategoriesEntity entityChanged = repository.save(entity);
@@ -38,11 +46,34 @@ public class CategoriesServiceImpl implements CategoriesService {
 
     @Override
     public void deleteCategory(Integer catId) {
+        if (repository.existsById(catId)) {
+            throw new DataNotFoundException("Category with id=" + catId + " was not found");
+        }
+        List<EventsEntity> eventsEntities = eventsRepository.findAllByCategoryId(catId);
+        if (eventsEntities.size() > 0) {
+            throw new EventValidationException("The category is not empty");
+        }
         CategoriesEntity entity = findCategoryById(catId);
         repository.delete(entity);
     }
 
     public CategoriesEntity findCategoryById(Integer categoryId) {
         return repository.findById(categoryId).get();
+    }
+
+    @Override
+    public List<CategoryDto> getCategories(Integer from, Integer size) {
+        Pageable pageParam = PageRequest.of(from > 0 ? from / size : 0, size);
+        List<CategoriesEntity> entities = repository.findAll(pageParam).toList();
+        return mapper.toDtoList(entities);
+    }
+
+    @Override
+    public CategoryDto getCategory(Integer catId) {
+        CategoriesEntity entity = repository.findById(catId).get();
+        if (entity == null) {
+            throw new DataNotFoundException("Category with id=" + catId + " was not found");
+        }
+        return mapper.toDto(entity);
     }
 }
